user_images_service/requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
asyncpg==0.24.0
click==8.0.1
cycler==0.10.0
databases==0.4.3
fastapi==0.68.0
h11==0.12.0
httptools==0.2.0
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
opencv-python-headless 
ormar==0.10.16
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.3.23
starlette==0.14.2
typing-extensions==3.7.4.3
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
rsa==4.7.2
python-jose==3.3.0
pyasn1==0.4.8

opencv-python==4.5.3.56
ecdsa==0.17.0
requests
